backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.15
psycopg2-binary>=2.9,<3.0
google-generativeai
django-cors-headers>=4.0,<5.0